refactor project structure and remove unused models
