+++ src/components/blocks/List.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import type Block from "../../models/block";
-  import FormattedText from "./FormattedText.svelte";
+  import FormattedText from "./components/FormattedText.svelte";
 
   export let block: Block;
   const isNumbered = block.content[0].type === "numbered_list";
