+++ src/components/blocks/Quote.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import type Block from "../../models/block";
-  import FormattedText from "../subBlocks/FormattedText.svelte";
+  import FormattedText from "./components/FormattedText.svelte";
 
   export let block: Block;
   const title = block.properties?.title ?? "";
