--- src/models/blocks/columnList.ts
@@ -1,5 +1,0 @@
-import Block from "../block";
-
-export default interface BlockColumnList extends Block {
-  content: BlockColumnList[];
-}
