--- src/models/blocks/list.ts
@@ -1,5 +1,0 @@
-import Block from "../block";
-
-export default interface BlockList extends Block {
-  content: Block[];
-}
